requirements/base.txt
# Base
Pillow==9.1.0
psycopg2==2.9.3

# Django
Django==4.0.4
django-cors-headers==3.12.0

# Environment
django-environ==0.8.1

# Django Rest Framework
djangorestframework==3.13.1
djangorestframework-gis==1.0

# JWT
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0

# Password hashing
argon2-cffi==21.3.0

dj_database_url

gunicorn==20.0.4

python-slugify==6.1.2
django-colorfield==0.7.2
django-filter==22.1
drf-spectacular==0.26.1

ffmpy==0.3.0
pyodk==0.1.0
pandas==1.5.3
django-webpack-loader==0.7.0

django-environ==0.8.1
ipython
Faker==18.10.1
